fix various comparision missmatches, missing sanity checks get rid of get_name2addr4() since it's unused refs #27

+++ src/common/cidr.h
@@ -34,7 +34,7 @@
 #define __CIDR_H__
 
 struct cidr_s {
-    unsigned long network;
+    u_int32_t network;
     int masklen;
     struct cidr_s *next;
 };
